Changed package resources name to models, to make clearer the arquitecture MVC we are following, also refactored everything necessary
