+++ app/src/go.mod
@@ -5,7 +5,6 @@ go 1.20
 require github.com/lib/pq v1.10.7
 
 require (
-	github.com/golang-migrate/migrate v3.5.4+incompatible
 	github.com/golang-migrate/migrate/v4 v4.15.2
 	github.com/hashicorp/errwrap v1.1.0 // indirect
 	github.com/hashicorp/go-multierror v1.1.1 // indirect
