+++ app/src/models/user.go
@@ -1,4 +1,4 @@
-package resources
+package models
 
 type User struct {
 	id    int
